requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
chromadb==0.4.18
sentence-transformers==2.2.2
networkx==3.2.1
spacy==3.7.2
pandas==2.1.3
requests==2.31.0
torch==2.1.1
dspy-ai==2.4.5
python-dotenv==1.0.0
pydantic==2.5.0 